feat: unify configuration
